requirements.txt
apache-airflow-providers-slack==8.4.0
apache-airflow-providers-snowflake==5.1.2
dash
kaleido==0.2.1
matplotlib==3.8.2
nbformat==5.9.2
plotly==5.18.0
python-dateutil==2.9.0
scikit-learn==1.3.2
statsmodels==0.14.0
wandb==0.16.3
